refactor(type): move to dedicate type file
